PRR-44 [fix]: function props drilling 관련 구조 일부 개선
