+++ src/constants/types/index.ts
@@ -2,3 +2,4 @@ export * from './apiTypes';
 export * from './authTypes';
 export * from './userTypes';
 export * from './projectTypes';
+export * from './facilityTypes';
