+++ src/pages/mypage/operation/apply.tsx
@@ -1,11 +1,8 @@
-import React, { useState, useRef } from 'react';
+import React, { useState } from 'react';
 
 import AddFacilityTemplate from '@/components/template/AddFacilityTemplate';
 
-import { AddFacilityInputType } from '@/constants/types/facilityTypes';
-import { CategoryType } from '~/src/constants/types';
-
-const MAX_FILE_SIZE = 3 * 1024 * 1024;
+import { AddFacilityInputType } from '@/constants/types';
 
 const Apply = () => {
   const [facilityInformation, setFacilityInformation] =
@@ -19,27 +16,6 @@ const Apply = () => {
       certificationDoc: null,
     });
 
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const handleFormInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name: inputName, value } = e.target;
-    setFacilityInformation((prev) => ({ ...prev, [inputName]: value }));
-    console.log('?');
-  };
-
-  const handleUploadFile = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const [uploadedFile] = e.target.files ?? [];
-    if (!uploadedFile || uploadedFile.size > MAX_FILE_SIZE) return;
-    setFacilityInformation((prev) => ({
-      ...prev,
-      certificationDoc: uploadedFile,
-    }));
-  };
-
-  const handleSelectCategory = (category: CategoryType) => {
-    setFacilityInformation((prev) => ({ ...prev, category }));
-  };
-
   return (
     <AddFacilityTemplate
       category={facilityInformation.category}
@@ -49,10 +25,7 @@ const Apply = () => {
       district={facilityInformation.district}
       detail={facilityInformation.detail}
       certificationDoc={facilityInformation.certificationDoc}
-      fileInputRef={fileInputRef}
-      handleFormInput={handleFormInput}
-      handleSelectCategory={handleSelectCategory}
-      handleUploadFile={handleUploadFile}
+      setFacilityInformation={setFacilityInformation}
     />
   );
 };
